refactor: cоздать необходимые react-компоненты
