+++ src/components/App/App.js
@@ -1,7 +1,7 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "../../images/logo.svg";
+import "./App.css";
 
-function App() {
+export default function App() {
   return (
     <div className="App">
       <header className="App-header">
@@ -21,5 +21,3 @@ function App() {
     </div>
   );
 }
-
-export default App;
